\name{mar1s-package}
\alias{mar1s-package}
\alias{mar1s}
\docType{package}
\title{
  Multiplicative AR(1) with Seasonal Processes
}
\description{
  Multiplicative AR(1) with Seasonal is a stochastic process model built
  on top of AR(1). The package provides the following procedures for
  MAR(1)S processes: fit, compose, decompose, advanced simulate and
  predict.
}
\details{
\tabular{ll}{
  Package: \tab mar1s\cr
  Type: \tab Package\cr
  Version: \tab 2.0-1\cr
  Date: \tab 2008-06-12\cr
  License: \tab GPL (>= 3)\cr
}
}
\author{
  Andrey Paramonov <cmr.Pent@gmail.com>
}
\keyword{ package }
\keyword{ models }
\keyword{ multivariate }
\keyword{ ts }
